Safari Park Inheritance Part 1

+++ WPF_Window/Calculator_Model/Calculator.cs
@@ -4,9 +4,8 @@ namespace Calculator_Model
 {
     public class Calculator
     {
-        public int Num1 { get; set; }
-
-        public int Num2 { get; set; }
+        public double Num1 { get; set; }
+        public double Num2 { get; set; }
 
         public double Add()
         {
